[sharktank] auto-unboxing and tree support in assert_tensor_close (#1797)
We want tensor comparison to accept unboxed and boxed tensors, so that
we don't have to do an explicit unboxing before a call. This makes the
function behave like the rest of our polymorphic operations.

Add support for tensor trees and rtol as arguments like
torch.testing.assert_close does.