[sharktank][quantization] remove redundant unpacking (#1836)
We are redundantly unpacking and returning a raw tensor when our whole
api exists to not have to do this.


EDIT:: Note to anyone reading this in the future because you want to
know why `probs_quantizer` was removed. It was implemented here
incompletely and incorrectly because it was modifying attention_scale. A
correct implementation would not interfere with the attn scale, as
`probs_quantizer` as implemented by quark only applies to the
post-softmax output in attention