[sharktank] propagate transpose_rhs into the shards on sharded matmumls (#1902)
The FP4 block quantized tensor does not support transposition of the
blocked dimension and it is not clear how to support this. This forces
us to not use transposition to simplify matmuls, but to propagate the
bool arg transpose_rhs to the multiplication of the shards.

Rework the tests to cover more cases.