Generating logs for export, compilation and iree_benchmark test (#2035)
Adding support for generating logs for export, compilation and
iree_benchmark test in model CI reports .

This will help adding links for the Failed tests in CI reports so that
anyone can directly see the logs if the test fails by clicking at the
links.

This will also help in extracting the time for each test.