[sharktank] in assert_tensor_close print abs diff and expected stats (#2124)
Currently when the assert fails we print just diff (actual - expected)
statistics, but for the error (diff) it is more meaningful what is the
absolute.

It is also informative to know the statistics of the expected. For
example the diff looses magnitude information, which is important when
judging the deviation.