[shortfin][llm] Move batching logic behind a strategy pattern (#2075)
This patch moves all the batching logic behind an opaque facade that is
exposed to the user.
Internal implementation details are completely hidden, and new batching
engines can be added through the use of a creation factory.
This ensures that major changes to the batching code do not touch other
parts of the code base, and it is easy to plug in new batching
algorithms.

Signed-off-by: Vinayak Dev <vinayak.dev@amd.com>
Co-authored-by: Stephen Baione
<109226581+stbaione@users.noreply.github.com>

---------

Signed-off-by: Vinayak Dev <vinayak.dev@amd.com>
Co-authored-by: Stephen Baione <stbaione@amd.com>