[Sharktank] Replumb Attention and embedding masks (#2132)
Instead of creating the masks outside of the model, passing them around,
and relying on IREE to fuse them for us, this PR changes it so that each
mask is created right before being used. This will make it easier to
fuse down the line and requires less state to be passed around.

---------

Signed-off-by: Alex Vasile <48962821+Alex-Vasile@users.noreply.github.com>